fix: change tv datafeed to hook for reset cache (#1131)
