+++ src/views/charts/TradingView/TvChart.tsx
@@ -1,5 +1,6 @@
 import { useEffect, useRef, useState } from 'react';
 
+import BigNumber from 'bignumber.js';
 import type { ResolutionString } from 'public/tradingview/charting_library';
 
 import { DEFAULT_MARKETID } from '@/constants/markets';
@@ -10,19 +11,26 @@ import { useChartLines } from '@/hooks/tradingView/useChartLines';
 import { useChartMarketAndResolution } from '@/hooks/tradingView/useChartMarketAndResolution';
 import { useOrderbookCandles } from '@/hooks/tradingView/useOrderbookCandles';
 import { useTradingView } from '@/hooks/tradingView/useTradingView';
+import { useTradingViewDatafeed } from '@/hooks/tradingView/useTradingViewDatafeed';
 import { useTradingViewTheme } from '@/hooks/tradingView/useTradingViewTheme';
 import { useTradingViewToggles } from '@/hooks/tradingView/useTradingViewToggles';
+import { useDydxClient } from '@/hooks/useDydxClient';
 import usePrevious from '@/hooks/usePrevious';
 
 import { useAppSelector } from '@/state/appTypes';
 import { getSelectedDisplayUnit } from '@/state/configsSelectors';
-import { getCurrentMarketId } from '@/state/perpetualsSelectors';
+import { getCurrentMarketConfig, getCurrentMarketId } from '@/state/perpetualsSelectors';
+
+import { orEmptyObj } from '@/lib/typeUtils';
 
 import { BaseTvChart } from './BaseTvChart';
 
 export const TvChart = () => {
   const [isChartReady, setIsChartReady] = useState(false);
   const currentMarketId: string = useAppSelector(getCurrentMarketId) ?? DEFAULT_MARKETID;
+  const { tickSizeDecimals: tickSizeDecimalsAbacus } = orEmptyObj(
+    useAppSelector(getCurrentMarketConfig)
+  );
 
   const tvWidgetRef = useRef<TvWidget | null>(null);
   const tvWidget = tvWidgetRef.current;
@@ -36,6 +44,31 @@ export const TvChart = () => {
   const buySellMarksToggleRef = useRef<HTMLElement | null>(null);
   const buySellMarksToggle = buySellMarksToggleRef.current;
 
+  const [tickSizeDecimalsIndexer, setTickSizeDecimalsIndexer] = useState<{
+    [marketId: string]: number | undefined;
+  }>({});
+
+  const tickSizeDecimals =
+    (currentMarketId
+      ? tickSizeDecimalsIndexer[currentMarketId] ?? tickSizeDecimalsAbacus
+      : tickSizeDecimalsAbacus) ?? undefined;
+
+  const { getMarketTickSize } = useDydxClient();
+
+  useEffect(() => {
+    // we only need tick size from current market for the price scale settings
+    // if markets haven't been loaded via abacus, get the current market info from indexer
+    (async () => {
+      if (currentMarketId && tickSizeDecimals === undefined) {
+        const marketTickSize = await getMarketTickSize(currentMarketId);
+        setTickSizeDecimalsIndexer((prev) => ({
+          ...prev,
+          [currentMarketId]: BigNumber(marketTickSize).decimalPlaces() ?? undefined,
+        }));
+      }
+    })();
+  }, [currentMarketId, getMarketTickSize, tickSizeDecimals]);
+
   const {
     orderLinesToggleOn,
     setOrderLinesToggleOn,
@@ -44,6 +77,12 @@ export const TvChart = () => {
     setBuySellMarksToggleOn,
     buySellMarksToggleOn,
   } = useTradingViewToggles();
+
+  const { datafeed, resetCache } = useTradingViewDatafeed({
+    orderbookCandlesToggleOn,
+    tickSizeDecimals,
+  });
+
   const { savedResolution } = useTradingView({
     tvWidgetRef,
     orderLineToggleRef,
@@ -56,6 +95,8 @@ export const TvChart = () => {
     buySellMarksToggleOn,
     setBuySellMarksToggleOn,
     setIsChartReady,
+    tickSizeDecimals,
+    datafeed,
   });
   useChartMarketAndResolution({
     currentMarketId,
@@ -92,9 +133,12 @@ export const TvChart = () => {
     // Only reset data if displayUnit has actually changed
     if (prevDisplayUnit !== displayUnit) {
       const chart = tvWidget.activeChart?.();
-      chart?.resetData();
+      if (chart) {
+        chart.resetData();
+        resetCache();
+      }
     }
-  }, [displayUnit, tvWidget, isChartReady, prevDisplayUnit]);
+  }, [displayUnit, tvWidget, isChartReady, prevDisplayUnit, resetCache]);
 
   return <BaseTvChart isChartReady={isChartReady} />;
 };
